refactor(architecture): Step 2 - Consolidate presentation layer (GenUI, streaming, scivis)
