+++ src/coreason_manifest/presentation/ambient.py
