+++ src/coreason_manifest/presentation/components/highlighting.py
