+++ src/coreason_manifest/presentation/components/search_layout.py
@@ -13,10 +13,10 @@ from enum import StrEnum
 from pydantic import Field, model_validator
 
 from coreason_manifest.core.common.base import CoreasonModel
-from coreason_manifest.core.presentation.highlighting import HighlightConfig
-from coreason_manifest.core.presentation.templating import ParameterizedDataRef
-from coreason_manifest.core.presentation.transform import DataTransformSchema
-from coreason_manifest.core.presentation.typeahead import TypeaheadConfig
+from coreason_manifest.presentation.components.highlighting import HighlightConfig
+from coreason_manifest.presentation.components.templating import ParameterizedDataRef
+from coreason_manifest.presentation.components.transform import DataTransformSchema
+from coreason_manifest.presentation.components.typeahead import TypeaheadConfig
 
 
 class PartitionType(StrEnum):
