+++ src/coreason_manifest/presentation/components/templating.py
