+++ src/coreason_manifest/presentation/components/visualizer.py
