+++ src/coreason_manifest/presentation/streaming.py
