+++ src/coreason_manifest/telemetry/suspense_envelope.py
@@ -2,7 +2,7 @@ from typing import Literal
 
 from pydantic import Field
 
-from coreason_manifest.core.common.suspense import SuspenseConfig
+from coreason_manifest.presentation.streaming import SuspenseConfig
 from coreason_manifest.state.events import EpistemicEvent
 from coreason_manifest.telemetry.stream_base import BaseEnvelope
 from coreason_manifest.telemetry.telemetry_schemas import AgentSignature, HardwareFingerprint
